Use __all__ to re-export the public API and have tests use that. Move Context to models.py.
