+++ src/wired/dataclasses/models.py
@@ -1,0 +1,3 @@
+class Context:
+    """ Marker interface for lookups by request context """
+    pass
